Splitting Code
Splitting code into a connection file, function file and index file.
This should allow for better readability.
